Merged patches from Fran Taylor.
Fran <tuneology@gmail.com> wrote:
Attached is a patch that I have written for dhcpd so that DHCPv6
client information can be stored in LDAP.

It is admittedly fairly crude, but I have been running it for over a
year now and it seems to work fine.  I have tested it with linux,
Solaris, Windows, and OSX clients.